update refactored core - tests fail on img and newline
