+++ TypeScript/source/core/common.ts
@@ -1,10 +1,13 @@
-/* © 2016-2018 FlowCrypt Limited. Limitations apply. Contact human@flowcrypt.com */
+/* ©️ 2016 - present FlowCrypt a.s. Limitations apply. Contact human@flowcrypt.com */
 
 'use strict';
 
-import { base64encode, base64decode } from '../platform/util.js';
+import { base64decode, base64encode } from '../platform/util.js';
 
 export type Dict<T> = { [key: string]: T; };
+export type UrlParam = string | number | null | undefined | boolean | string[];
+export type UrlParams = Dict<UrlParam>;
+export type PromiseCancellation = { cancel: boolean };
 
 export class Str {
 
@@ -23,13 +26,30 @@ export class Str {
     return { email, name, full };
   }
 
-  public static prettyPrint = (obj: any) => (typeof obj === 'object') ? JSON.stringify(obj, undefined, 2).replace(/ /g, '&nbsp;').replace(/\n/g, '<br />') : String(obj);
+  public static rmSpecialCharsKeepUtf = (str: string, mode: 'ALLOW-SOME' | 'ALLOW-NONE'): string => {
+    // not a whitelist because we still want utf chars
+    str = str.replace(/[@&#`();:'",<>\{\}\[\]\\\/\n\t\r]/gi, '');
+    if (mode === 'ALLOW-SOME') {
+      return str;
+    }
+    return str.replace(/[.~!$%^*=?]/gi, '');
+  }
 
-  public static normalizeSpaces = (str: string) => str.replace(RegExp(String.fromCharCode(160), 'g'), String.fromCharCode(32));
+  public static prettyPrint = (obj: any) => {
+    return (typeof obj === 'object') ? JSON.stringify(obj, undefined, 2).replace(/ /g, '&nbsp;').replace(/\n/g, '<br />') : String(obj);
+  }
 
-  public static normalizeDashes = (str: string) => str.replace(/^—–|—–$/gm, '-----');
+  public static normalizeSpaces = (str: string) => {
+    return str.replace(RegExp(String.fromCharCode(160), 'g'), String.fromCharCode(32));
+  }
 
-  public static normalize = (str: string) => Str.normalizeSpaces(Str.normalizeDashes(str));
+  public static normalizeDashes = (str: string) => {
+    return str.replace(/^—–|—–$/gm, '-----');
+  }
+
+  public static normalize = (str: string) => {
+    return Str.normalizeSpaces(Str.normalizeDashes(str));
+  }
 
   public static numberFormat = (number: number) => {
     const nStr: string = number + '';
@@ -50,7 +70,9 @@ export class Str {
     return /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/i.test(email);
   }
 
-  public static monthName = (monthIndex: number) => ['January', 'February', 'March', 'April', 'May', 'June', 'July', 'August', 'September', 'October', 'November', 'December'][monthIndex];
+  public static monthName = (monthIndex: number) => {
+    return ['January', 'February', 'March', 'April', 'May', 'June', 'July', 'August', 'September', 'October', 'November', 'December'][monthIndex];
+  }
 
   public static sloppyRandom = (length: number = 5) => {
     let id = '';
@@ -61,13 +83,17 @@ export class Str {
     return id;
   }
 
-  public static regexEscape = (toBeUsedInRegex: string) => toBeUsedInRegex.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
+  public static regexEscape = (toBeUsedInRegex: string) => {
+    return toBeUsedInRegex.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
+  }
 
   public static asEscapedHtml = (text: string) => {
     return text.replace(/&/g, '&amp;').replace(/"/g, '&quot;').replace(/'/g, '&#39;').replace(/</g, '&lt;').replace(/>/g, '&gt;').replace(/\//g, '&#x2F;').replace(/\n/g, '<br />');
   }
 
-  public static htmlAttrEncode = (values: Dict<any>): string => Str.base64urlUtfEncode(JSON.stringify(values));
+  public static htmlAttrEncode = (values: Dict<any>): string => {
+    return Str.base64urlUtfEncode(JSON.stringify(values));
+  }
 
   public static htmlAttrDecode = (encoded: string): any => {
     try {
@@ -77,13 +103,21 @@ export class Str {
     }
   }
 
-  public static capitalize = (string: string): string => string.trim().split(' ').map(s => s.charAt(0).toUpperCase() + s.slice(1)).join(' ');
+  public static capitalize = (string: string): string => {
+    return string.trim().split(' ').map(s => s.charAt(0).toUpperCase() + s.slice(1)).join(' ');
+  }
 
-  public static toUtcTimestamp = (datetimeStr: string, asStr: boolean = false) => asStr ? String(Date.parse(datetimeStr)) : Date.parse(datetimeStr);
+  public static toUtcTimestamp = (datetimeStr: string, asStr: boolean = false) => {
+    return asStr ? String(Date.parse(datetimeStr)) : Date.parse(datetimeStr);
+  }
 
-  public static datetimeToDate = (date: string) => date.substr(0, 10).replace(/"/g, '&quot;').replace(/'/g, '&#39;').replace(/</g, '&lt;');
+  public static datetimeToDate = (date: string) => {
+    return date.substr(0, 10).replace(/"/g, '&quot;').replace(/'/g, '&#39;').replace(/</g, '&lt;');
+  }
 
-  public static fromDate = (date: Date) => date.toISOString().replace(/T/, ' ').replace(/:[^:]+$/, '');
+  public static fromDate = (date: Date) => {
+    return date.toISOString().replace(/T/, ' ').replace(/:[^:]+$/, '');
+  }
 
   private static base64urlUtfEncode = (str: string) => {
     // https://stackoverflow.com/questions/30106476/using-javascripts-atob-to-decode-base64-doesnt-properly-decode-utf-8-strings
@@ -153,5 +187,88 @@ export class Value {
   };
 
   public static noop = (): void => undefined;
+
+}
+
+export class Url {
+
+  private static URL_PARAM_DICT: Dict<boolean | null> = { '___cu_true___': true, '___cu_false___': false, '___cu_null___': null }; // tslint:disable-line:no-null-keyword
+
+  /**
+   * will convert result to desired format: camelCase or snake_case, based on what was supplied in expectedKeys
+   * todo - the camelCase or snake_case functionality can now be removed
+   */
+  public static parse = (expectedKeys: string[], parseThisUrl?: string) => {
+    const url = (parseThisUrl || window.location.search.replace('?', ''));
+    const valuePairs = url.split('?').pop()!.split('&'); // str.split('?') string[].length will always be >= 1
+    const rawParams: Dict<string> = {};
+    const rawParamNameDict: Dict<string> = {};
+    for (const valuePair of valuePairs) {
+      const pair = valuePair.split('=');
+      rawParams[pair[0]] = pair[1];
+      Url.fillPossibleUrlParamNameVariations(pair[0], rawParamNameDict);
+    }
+    const processedParams: UrlParams = {};
+    for (const expectedKey of expectedKeys) {
+      processedParams[expectedKey] = Url.findAndProcessUrlParam(expectedKey, rawParamNameDict, rawParams);
+    }
+    return processedParams;
+  }
+
+  public static create = (link: string, params: UrlParams) => {
+    for (const key of Object.keys(params)) {
+      const value = params[key];
+      if (typeof value !== 'undefined') {
+        const transformed = Value.obj.keyByValue(Url.URL_PARAM_DICT, value);
+        link += (link.includes('?') ? '&' : '?') + encodeURIComponent(key) + '=' + encodeURIComponent(String(typeof transformed !== 'undefined' ? transformed : value));
+      }
+    }
+    return link;
+  }
+
+  public static removeParamsFromUrl = (url: string, paramsToDelete: string[]) => {
+    const urlParts = url.split('?');
+    if (!urlParts[1]) { // Nothing to remove
+      return url;
+    }
+    let queryParams = urlParts[1];
+    queryParams = queryParams[queryParams.length - 1] === '#' ? queryParams.slice(0, -1) : queryParams;
+    const params = new URLSearchParams(queryParams);
+    for (const p of paramsToDelete) {
+      params.delete(p);
+    }
+    return `${urlParts[0]}?${params.toString()}`;
+  }
+
+  private static snakeCaseToCamelCase = (s: string) => {
+    return s.replace(/_[a-z]/g, boundary => boundary[1].toUpperCase());
+  }
+
+  private static camelCaseToSnakeCase = (s: string) => {
+    return s.replace(/[a-z][A-Z]/g, boundary => `${boundary[0]}_${boundary[1].toLowerCase()}`);
+  }
+
+  private static findAndProcessUrlParam = (expectedParamName: string, rawParamNameDict: Dict<string>, rawParms: Dict<string>): UrlParam => {
+    if (typeof rawParamNameDict[expectedParamName] === 'undefined') {
+      return undefined; // param name not found in param name dict
+    }
+    const rawValue = rawParms[rawParamNameDict[expectedParamName]];
+    if (typeof rawValue === 'undefined') {
+      return undefined; // original param name not found in raw params
+    }
+    if (typeof Url.URL_PARAM_DICT[rawValue] !== 'undefined') {
+      return Url.URL_PARAM_DICT[rawValue]; // raw value was converted using a value dict to get proper: true, false, undefined, null
+    }
+    return decodeURIComponent(rawValue);
+  }
+
+  private static fillPossibleUrlParamNameVariations = (urlParamName: string, rawParamNameDict: Dict<string>) => {
+    rawParamNameDict[urlParamName] = urlParamName;
+    rawParamNameDict[Url.snakeCaseToCamelCase(urlParamName)] = urlParamName;
+    rawParamNameDict[Url.camelCaseToSnakeCase(urlParamName)] = urlParamName;
+    const shortened = urlParamName.replace('account', 'acct').replace('message', 'msg').replace('attachment', 'att');
+    rawParamNameDict[Url.snakeCaseToCamelCase(shortened)] = urlParamName;
+    rawParamNameDict[Url.camelCaseToSnakeCase(shortened)] = urlParamName;
+  }
 
 }
