+++ TypeScript/source/core/const.ts
@@ -1,4 +1,4 @@
-/* © 2016-2018 FlowCrypt Limited. Limitations apply. Contact human@flowcrypt.com */
+/* ©️ 2016 - present FlowCrypt a.s. Limitations apply. Contact human@flowcrypt.com */
 
 'use strict';
 
@@ -6,6 +6,8 @@ export const VERSION = '[BUILD_REPLACEABLE_VERSION]';
 export const GOOGLE_API_HOST = '[BUILD_REPLACEABLE_GOOGLE_API_HOST]';
 export const GOOGLE_OAUTH_SCREEN_HOST = '[BUILD_REPLACEABLE_GOOGLE_OAUTH_SCREEN_HOST]';
 export const GOOGLE_CONTACTS_API_HOST = '[BUILD_REPLACEABLE_GOOGLE_CONTACTS_API_HOST]';
+export const BACKEND_API_HOST = '[BUILD_REPLACEABLE_BACKEND_API_HOST]';
+
 /**
  * Only put constants below if:
  *   - they are useful across web/extension/Nodejs environments, AND
