+++ TypeScript/source/core/mime.ts
@@ -1,14 +1,18 @@
-/* © 2016-2018 FlowCrypt Limited. Limitations apply. Contact human@flowcrypt.com */
+/* ©️ 2016 - present FlowCrypt a.s. Limitations apply. Contact human@flowcrypt.com */
 
 'use strict';
 
-import { Str, Dict } from './common.js';
-import { Pgp, KeyDetails, DecryptError, VerifyRes } from './pgp.js';
-import { Att, AttMeta } from './att.js';
-import { Catch } from '../platform/catch.js';
-import { requireMimeParser, requireMimeBuilder, requireIso88592 } from '../platform/require.js';
+import { Dict, Str } from './common.js';
+import { requireIso88592, requireMimeBuilder, requireMimeParser } from '../platform/require.js';
+
+import { Att } from './att.js';
 import { Buf } from './buf.js';
+import { Catch } from '../platform/catch.js';
 import { MimeParserNode } from './types/emailjs';
+import { MsgBlock } from './msg-block.js';
+import { MsgBlockParser } from './msg-block-parser.js';
+import { PgpArmor } from './pgp-armor.js';
+import { iso2022jpToUtf } from '../platform/util.js';
 
 const MimeParser = requireMimeParser();  // tslint:disable-line:variable-name
 const MimeBuilder = requireMimeBuilder();  // tslint:disable-line:variable-name
@@ -30,22 +34,9 @@ export type MimeContent = {
   bcc: string[];
 };
 
+export type MimeEncodeType = 'pgpMimeEncrypted' | 'pgpMimeSigned' | undefined;
 export type RichHeaders = Dict<string | string[]>;
 export type SendableMsgBody = { [key: string]: string | undefined; 'text/plain'?: string; 'text/html'?: string; };
-export type KeyBlockType = 'publicKey' | 'privateKey';
-export type ReplaceableMsgBlockType = KeyBlockType | 'cryptupVerification' | 'signedMsg' | 'encryptedMsg' | 'encryptedMsgLink';
-export type MsgBlockType = ReplaceableMsgBlockType | 'plainText' | 'decryptedText' | 'plainHtml' | 'decryptedHtml' | 'plainAtt' | 'encryptedAtt'
-  | 'decryptedAtt' | 'encryptedAttLink' | 'decryptErr' | 'verifiedMsg' | 'signedHtml';
-export type MsgBlock = {
-  type: MsgBlockType;
-  content: string | Buf;
-  complete: boolean;
-  signature?: string;
-  keyDetails?: KeyDetails; // only in publicKey when returned to Android (could eventually be made mandatory, done straight in detectBlocks?)
-  attMeta?: AttMeta; // only in plainAtt, encryptedAtt, decryptedAtt, encryptedAttLink (not sure if always)
-  decryptErr?: DecryptError; // only in decryptErr block, always
-  verifyRes?: VerifyRes,
-};
 export type MimeProccesedMsg = {
   rawSignedContent: string | undefined,
   headers: Dict<MimeContentHeader>,
@@ -60,35 +51,35 @@ export class Mime {
   public static processDecoded = (decoded: MimeContent): MimeProccesedMsg => {
     const blocks: MsgBlock[] = [];
     if (decoded.text) {
-      const blocksFromTextPart = Pgp.armor.detectBlocks(Str.normalize(decoded.text)).blocks;
+      const blocksFromTextPart = MsgBlockParser.detectBlocks(Str.normalize(decoded.text)).blocks;
       // if there are some encryption-related blocks found in the text section, which we can use, and not look at the html section
-      if (blocksFromTextPart.find(b => b.type === 'encryptedMsg' || b.type === 'signedMsg' || b.type === 'publicKey' || b.type === 'privateKey' || b.type === 'cryptupVerification')) {
+      if (blocksFromTextPart.find(b => b.type === 'encryptedMsg' || b.type === 'signedMsg' || b.type === 'publicKey' || b.type === 'privateKey')) {
         blocks.push(...blocksFromTextPart); // because the html most likely containt the same thing, just harder to parse pgp sections cause it's html
       } else if (decoded.html) { // if no pgp blocks found in text part and there is html part, prefer html
-        blocks.push(Pgp.internal.msgBlockObj('plainHtml', decoded.html));
+        blocks.push(MsgBlock.fromContent('plainHtml', decoded.html));
       } else { // else if no html and just a plain text message, use that
         blocks.push(...blocksFromTextPart);
       }
     } else if (decoded.html) {
-      blocks.push(Pgp.internal.msgBlockObj('plainHtml', decoded.html));
+      blocks.push(MsgBlock.fromContent('plainHtml', decoded.html));
     }
     for (const file of decoded.atts) {
       const treatAs = file.treatAs();
       if (treatAs === 'encryptedMsg') {
-        const armored = Pgp.armor.clip(file.getData().toUtfStr());
+        const armored = PgpArmor.clip(file.getData().toUtfStr());
         if (armored) {
-          blocks.push(Pgp.internal.msgBlockObj('encryptedMsg', armored));
+          blocks.push(MsgBlock.fromContent('encryptedMsg', armored));
         }
       } else if (treatAs === 'signature') {
         decoded.signature = decoded.signature || file.getData().toUtfStr();
       } else if (treatAs === 'publicKey') {
-        blocks.push(...Pgp.armor.detectBlocks(file.getData().toUtfStr()).blocks);
+        blocks.push(...MsgBlockParser.detectBlocks(file.getData().toUtfStr()).blocks);
       } else if (treatAs === 'privateKey') {
-        blocks.push(...Pgp.armor.detectBlocks(file.getData().toUtfStr()).blocks);
+        blocks.push(...MsgBlockParser.detectBlocks(file.getData().toUtfStr()).blocks);
       } else if (treatAs === 'encryptedFile') {
-        blocks.push(Pgp.internal.msgBlockAttObj('encryptedAtt', '', { name: file.name, type: file.type, length: file.getData().length, data: file.getData() }));
+        blocks.push(MsgBlock.fromAtt('encryptedAtt', '', { name: file.name, type: file.type, length: file.getData().length, data: file.getData() }));
       } else if (treatAs === 'plainFile') {
-        blocks.push(Pgp.internal.msgBlockAttObj('plainAtt', '', {
+        blocks.push(MsgBlock.fromAtt('plainAtt', '', {
           name: file.name, type: file.type, length: file.getData().length, data: file.getData(), inline: file.inline, cid: file.cid
         }));
       }
@@ -104,7 +95,7 @@ export class Mime {
         }
       }
       if (!blocks.find(block => block.type === 'plainText' || block.type === 'plainHtml' || block.type === 'signedMsg' || block.type === 'signedHtml')) { // signed an empty message
-        blocks.push({ type: "signedMsg", "content": "", signature: decoded.signature, complete: true });
+        blocks.push(new MsgBlock("signedMsg", "", true, decoded.signature));
       }
     }
     return { headers: decoded.headers, blocks, from: decoded.from, to: decoded.to, rawSignedContent: decoded.rawSignedContent };
@@ -116,25 +107,7 @@ export class Mime {
   }
 
   public static isPlainInlineImg = (b: MsgBlock) => {
-    return b.type === 'plainAtt' && b.attMeta && b.attMeta.inline && b.attMeta.type && ['image/jpeg', 'image/jpg', 'image/bmp', 'image/png', 'image/svg+xml'].includes(b.attMeta.type);
-  }
-
-  private static headerGetAddress = (parsedMimeMsg: MimeContent, headersNames: Array<SendingType | 'from'>) => {
-    const result: { to: string[], cc: string[], bcc: string[] } = { to: [], cc: [], bcc: [] };
-    let from: string | undefined;
-    const getHdrValAsArr = (hdr: MimeContentHeader) => typeof hdr === 'string' ? [hdr].map(h => Str.parseEmail(h).email).filter(e => !!e) as string[] : hdr.map(h => h.address);
-    const getHdrValAsStr = (hdr: MimeContentHeader) => Str.parseEmail((Array.isArray(hdr) ? (hdr[0] || {}).address : String(hdr || '')) || '').email;
-    for (const hdrName of headersNames) {
-      const header = parsedMimeMsg.headers[hdrName];
-      if (header) {
-        if (hdrName === 'from') {
-          from = getHdrValAsStr(header);
-        } else {
-          result[hdrName] = [...result[hdrName], ...getHdrValAsArr(header)];
-        }
-      }
-    }
-    return { ...result, from };
+    return b.type === 'plainAtt' && b.attMeta?.inline && b.attMeta.type && ['image/jpeg', 'image/jpg', 'image/bmp', 'image/png', 'image/svg+xml'].includes(b.attMeta.type);
   }
 
   public static replyHeaders = (parsedMimeMsg: MimeContent) => {
@@ -153,6 +126,139 @@ export class Mime {
       return true;
     }
     return Boolean(contentType.index === 0 && utf8.match(/boundary=/));
+  }
+
+  public static decode = async (mimeMsg: Uint8Array): Promise<MimeContent> => {
+    let mimeContent: MimeContent = { atts: [], headers: {}, subject: undefined, text: undefined, html: undefined, signature: undefined, from: undefined, to: [], cc: [], bcc: [] };
+    const parser = new MimeParser();
+    const leafNodes: { [key: string]: MimeParserNode } = {};
+    parser.onbody = (node: MimeParserNode) => {
+      const path = String(node.path.join('.'));
+      if (typeof leafNodes[path] === 'undefined') {
+        leafNodes[path] = node;
+      }
+    };
+    return await new Promise((resolve, reject) => {
+      try {
+        parser.onend = () => {
+          try {
+            for (const name of Object.keys(parser.node.headers)) {
+              mimeContent.headers[name] = parser.node.headers[name][0].value;
+            }
+            mimeContent.rawSignedContent = Mime.retrieveRawSignedContent([parser.node]);
+            for (const node of Object.values(leafNodes)) {
+              if (Mime.getNodeType(node) === 'application/pgp-signature') {
+                mimeContent.signature = node.rawContent;
+              } else if (Mime.getNodeType(node) === 'text/html' && !Mime.getNodeFilename(node)) {
+                // html content may be broken up into smaller pieces by attachments in between
+                // AppleMail does this with inline attachments
+                mimeContent.html = (mimeContent.html || '') + Mime.getNodeContentAsUtfStr(node);
+              } else if (Mime.getNodeType(node) === 'text/plain' && !Mime.getNodeFilename(node)) {
+                mimeContent.text = Mime.getNodeContentAsUtfStr(node);
+              } else if (Mime.getNodeType(node) === 'text/rfc822-headers') {
+                if (node._parentNode && node._parentNode.headers.subject) {
+                  mimeContent.subject = node._parentNode.headers.subject[0].value;
+                }
+              } else {
+                mimeContent.atts.push(Mime.getNodeAsAtt(node));
+              }
+            }
+            const headers = Mime.headerGetAddress(mimeContent, ['from', 'to', 'cc', 'bcc']);
+            mimeContent.subject = String(mimeContent.subject || mimeContent.headers.subject || '(no subject)');
+            mimeContent = Object.assign(mimeContent, headers);
+            resolve(mimeContent);
+          } catch (e) {
+            reject(e);
+          }
+        };
+        parser.write(mimeMsg);
+        parser.end();
+      } catch (e) { // todo - on Android we may want to fail when this happens, evaluate effect on browser extension
+        Catch.reportErr(e);
+        resolve(mimeContent);
+      }
+    });
+  }
+
+  public static encode = async (body: SendableMsgBody, headers: RichHeaders, atts: Att[] = [], type?: MimeEncodeType): Promise<string> => {
+    const rootContentType = type !== 'pgpMimeEncrypted' ? 'multipart/mixed' : `multipart/encrypted; protocol="application/pgp-encrypted";`;
+    const rootNode = new MimeBuilder(rootContentType, { includeBccInHeader: true }); // tslint:disable-line:no-unsafe-any
+    for (const key of Object.keys(headers)) {
+      rootNode.addHeader(key, headers[key]); // tslint:disable-line:no-unsafe-any
+    }
+    if (Object.keys(body).length) {
+      let contentNode: MimeParserNode;
+      if (Object.keys(body).length === 1) {
+        contentNode = Mime.newContentNode(MimeBuilder, Object.keys(body)[0], body[Object.keys(body)[0] as "text/plain" | "text/html"] || '');
+      } else {
+        contentNode = new MimeBuilder('multipart/alternative'); // tslint:disable-line:no-unsafe-any
+        for (const type of Object.keys(body)) {
+          contentNode.appendChild(Mime.newContentNode(MimeBuilder, type, body[type]!)); // already present, that's why part of for loop
+        }
+      }
+      rootNode.appendChild(contentNode); // tslint:disable-line:no-unsafe-any
+    }
+    for (const att of atts) {
+      rootNode.appendChild(Mime.createAttNode(att)); // tslint:disable-line:no-unsafe-any
+    }
+    return rootNode.build(); // tslint:disable-line:no-unsafe-any
+  }
+
+  public static subjectWithoutPrefixes = (subject: string): string => {
+    return subject.replace(/^((Re|Fwd): ?)+/g, '').trim();
+  }
+
+  public static encodePgpMimeSigned = async (body: SendableMsgBody, headers: RichHeaders, atts: Att[] = [], sign: (data: string) => Promise<string>): Promise<string> => {
+    const sigPlaceholder = `SIG_PLACEHOLDER_${Str.sloppyRandom(10)}`;
+    const rootNode = new MimeBuilder(`multipart/signed; protocol="application/pgp-signature";`, { includeBccInHeader: true }); // tslint:disable-line:no-unsafe-any
+    for (const key of Object.keys(headers)) {
+      rootNode.addHeader(key, headers[key]); // tslint:disable-line:no-unsafe-any
+    }
+    const bodyNodes = new MimeBuilder('multipart/alternative'); // tslint:disable-line:no-unsafe-any
+    for (const type of Object.keys(body)) {
+      bodyNodes.appendChild(Mime.newContentNode(MimeBuilder, type, body[type]!)); // tslint:disable-line:no-unsafe-any
+    }
+    const signedContentNode = new MimeBuilder('multipart/mixed'); // tslint:disable-line:no-unsafe-any
+    signedContentNode.appendChild(bodyNodes); // tslint:disable-line:no-unsafe-any
+    for (const att of atts) {
+      signedContentNode.appendChild(Mime.createAttNode(att)); // tslint:disable-line:no-unsafe-any
+    }
+    const sigAttPlaceholder = new Att({ data: Buf.fromUtfStr(sigPlaceholder), type: 'application/pgp-signature', name: 'signature.asc' });
+    const sigAttPlaceholderNode = Mime.createAttNode(sigAttPlaceholder); // tslint:disable-line:no-unsafe-any
+    // https://tools.ietf.org/html/rfc3156#section-5 - signed content first, signature after
+    rootNode.appendChild(signedContentNode); // tslint:disable-line:no-unsafe-any
+    rootNode.appendChild(sigAttPlaceholderNode); // tslint:disable-line:no-unsafe-any
+    const mimeStrWithPlaceholderSig = rootNode.build() as string; // tslint:disable-line:no-unsafe-any
+    const { rawSignedContent } = await Mime.decode(Buf.fromUtfStr(mimeStrWithPlaceholderSig));
+    if (!rawSignedContent) {
+      console.log(`mimeStrWithPlaceholderSig(placeholder:${sigPlaceholder}):\n${mimeStrWithPlaceholderSig}`);
+      throw new Error('Could not find raw signed content immediately after mime-encoding a signed message');
+    }
+    const realSignature = await sign(rawSignedContent); // tslint:disable-line:no-unsafe-any
+    const pgpMimeSigned = mimeStrWithPlaceholderSig.replace(Buf.fromUtfStr(sigPlaceholder).toBase64Str(), Buf.fromUtfStr(realSignature).toBase64Str());
+    if (pgpMimeSigned === mimeStrWithPlaceholderSig) {
+      console.log(`pgpMimeSigned(placeholder:${sigPlaceholder}):\n${pgpMimeSigned}`);
+      throw new Error('Replaced sigPlaceholder with realSignature but mime stayed the same');
+    }
+    return pgpMimeSigned;
+  }
+
+  private static headerGetAddress = (parsedMimeMsg: MimeContent, headersNames: Array<SendingType | 'from'>) => {
+    const result: { to: string[], cc: string[], bcc: string[] } = { to: [], cc: [], bcc: [] };
+    let from: string | undefined;
+    const getHdrValAsArr = (hdr: MimeContentHeader) => typeof hdr === 'string' ? [hdr].map(h => Str.parseEmail(h).email).filter(e => !!e) as string[] : hdr.map(h => h.address);
+    const getHdrValAsStr = (hdr: MimeContentHeader) => Str.parseEmail((Array.isArray(hdr) ? (hdr[0] || {}).address : String(hdr || '')) || '').email;
+    for (const hdrName of headersNames) {
+      const header = parsedMimeMsg.headers[hdrName];
+      if (header) {
+        if (hdrName === 'from') {
+          from = getHdrValAsStr(header);
+        } else {
+          result[hdrName] = [...result[hdrName], ...getHdrValAsArr(header)];
+        }
+      }
+    }
+    return { ...result, from };
   }
 
   private static retrieveRawSignedContent = (nodes: MimeParserNode[]): string | undefined => {
@@ -176,84 +282,23 @@ export class Mime {
     return undefined;
   }
 
-  public static decode = (mimeMsg: Uint8Array): Promise<MimeContent> => {
-    return new Promise(async resolve => {
-      let mimeContent: MimeContent = { atts: [], headers: {}, subject: undefined, text: undefined, html: undefined, signature: undefined, from: undefined, to: [], cc: [], bcc: [] };
-      try {
-        const parser = new MimeParser();
-        const leafNodes: { [key: string]: MimeParserNode } = {};
-        parser.onbody = (node: MimeParserNode) => {
-          const path = String(node.path.join('.'));
-          if (typeof leafNodes[path] === 'undefined') {
-            leafNodes[path] = node;
-          }
-        };
-        parser.onend = () => {
-          for (const name of Object.keys(parser.node.headers)) {
-            mimeContent.headers[name] = parser.node.headers[name][0].value;
-          }
-          mimeContent.rawSignedContent = Mime.retrieveRawSignedContent([parser.node]);
-          for (const node of Object.values(leafNodes)) {
-            if (Mime.getNodeType(node) === 'application/pgp-signature') {
-              mimeContent.signature = node.rawContent;
-            } else if (Mime.getNodeType(node) === 'text/html' && !Mime.getNodeFilename(node)) {
-              // html content may be broken up into smaller pieces by attachments in between
-              // AppleMail does this with inline attachments
-              mimeContent.html = (mimeContent.html || '') + Mime.getNodeContentAsUtfStr(node);
-            } else if (Mime.getNodeType(node) === 'text/plain' && !Mime.getNodeFilename(node)) {
-              mimeContent.text = Mime.getNodeContentAsUtfStr(node);
-            } else if (Mime.getNodeType(node) === 'text/rfc822-headers') {
-              if (node._parentNode && node._parentNode.headers.subject) {
-                mimeContent.subject = node._parentNode.headers.subject[0].value;
-              }
-            } else {
-              mimeContent.atts.push(Mime.getNodeAsAtt(node));
-            }
-          }
-          const headers = Mime.headerGetAddress(mimeContent, ['from', 'to', 'cc', 'bcc']);
-          mimeContent.subject = String(mimeContent.subject || mimeContent.headers.subject || '(no subject)');
-          mimeContent = Object.assign(mimeContent, headers);
-          resolve(mimeContent);
-        };
-        parser.write(mimeMsg);
-        parser.end();
-      } catch (e) { // todo - on Android we may want to fail when this happens, evaluate effect on browser extension
-        Catch.reportErr(e);
-        resolve(mimeContent);
-      }
-    });
+  private static createAttNode = (att: Att): any => { // todo: MimeBuilder types
+    const type = `${att.type}; name="${att.name}"`;
+    const id = `f_${Str.sloppyRandom(30)}@flowcrypt`;
+    const header: Dict<string> = {};
+    if (att.contentDescription) {
+      header['Content-Description'] = att.contentDescription;
+    }
+    header['Content-Disposition'] = att.inline ? 'inline' : 'attachment';
+    header['X-Attachment-Id'] = id;
+    header['Content-ID'] = `<${id}>`;
+    header['Content-Transfer-Encoding'] = 'base64';
+    return new MimeBuilder(type, { filename: att.name }).setHeader(header).setContent(att.getData()); // tslint:disable-line:no-unsafe-any
   }
 
-  public static encode = async (body: string | SendableMsgBody, headers: RichHeaders, atts: Att[] = []): Promise<string> => {
-    const rootNode = new MimeBuilder('multipart/mixed', { includeBccInHeader: true }); // tslint:disable-line:no-unsafe-any
-    for (const key of Object.keys(headers)) {
-      rootNode.addHeader(key, headers[key]); // tslint:disable-line:no-unsafe-any
-    }
-    if (typeof body === 'string') {
-      body = { 'text/plain': body };
-    }
-    let contentNode: MimeParserNode;
-    if (Object.keys(body).length === 1) {
-      contentNode = Mime.newContentNode(MimeBuilder, Object.keys(body)[0], body[Object.keys(body)[0] as "text/plain" | "text/html"] || '');
-    } else {
-      contentNode = new MimeBuilder('multipart/alternative'); // tslint:disable-line:no-unsafe-any
-      for (const type of Object.keys(body)) {
-        contentNode.appendChild(Mime.newContentNode(MimeBuilder, type, body[type]!)); // already present, that's why part of for loop
-      }
-    }
-    rootNode.appendChild(contentNode); // tslint:disable-line:no-unsafe-any
-    for (const att of atts) {
-      const type = `${att.type}; name="${att.name}"`;
-      const id = `f_${Str.sloppyRandom(30)}@flowcrypt`;
-      const header = { 'Content-Disposition': 'attachment', 'X-Attachment-Id': id, 'Content-ID': `<${id}>`, 'Content-Transfer-Encoding': 'base64' };
-      rootNode.appendChild(new MimeBuilder(type, { filename: att.name }).setHeader(header).setContent(att.getData())); // tslint:disable-line:no-unsafe-any
-    }
-    return rootNode.build(); // tslint:disable-line:no-unsafe-any
-  }
-
-  private static getNodeType = (node: MimeParserNode) => {
+  private static getNodeType = (node: MimeParserNode, type: 'value' | 'initial' = 'value') => {
     if (node.headers['content-type'] && node.headers['content-type'][0]) {
-      return node.headers['content-type'][0].value;
+      return node.headers['content-type'][0][type];
     }
     return undefined;
   }
@@ -298,16 +343,21 @@ export class Mime {
   }
 
   private static getNodeContentAsUtfStr = (node: MimeParserNode): string => {
-    if (node.charset === 'utf-8' && node.contentTransferEncoding.value === 'base64') {
-      return Buf.fromUint8(node.content).toUtfStr();
-    }
-    if (node.charset === 'utf-8' && node.contentTransferEncoding.value === 'quoted-printable') {
-      return Mime.fromEqualSignNotationAsBuf(node.rawContent!).toUtfStr();
-    }
     if (node.charset && Iso88592.labels.includes(node.charset)) {
       return Iso88592.decode(node.rawContent!); // tslint:disable-line:no-unsafe-any
     }
-    return Buf.fromRawBytesStr(node.rawContent!).toUtfStr();
+    let resultBuf: Buf;
+    if (node.charset === 'utf-8' && node.contentTransferEncoding.value === 'base64') {
+      resultBuf = Buf.fromUint8(node.content);
+    } else if (node.charset === 'utf-8' && node.contentTransferEncoding.value === 'quoted-printable') {
+      resultBuf = Mime.fromEqualSignNotationAsBuf(node.rawContent!);
+    } else {
+      resultBuf = Buf.fromRawBytesStr(node.rawContent!);
+    }
+    if (node.charset?.toUpperCase() === 'ISO-2022-JP' || (node.charset === 'utf-8' && Mime.getNodeType(node, 'initial')?.includes('ISO-2022-JP'))) {
+      return iso2022jpToUtf(resultBuf);
+    }
+    return resultBuf.toUtfStr();
   }
 
   // tslint:disable-next-line:variable-name
