+++ TypeScript/source/platform/util.ts
@@ -2,6 +2,7 @@
 
 'use strict';
 
+import { Buf } from '../core/buf';
 import { randomBytes } from 'crypto';
 
 export const secureRandomBytes = (length: number): Uint8Array => {
@@ -20,3 +21,7 @@ export const setGlobals = () => {
   (global as any).btoa = base64encode;
   (global as any).atob = base64decode;
 }
+
+export const iso2022jpToUtf = (content: Buf) => {
+  throw new Error('iso2022jpToUtf not implemented on node');
+};
